refact: useScroll hook 리팩토링
- 범용성을 위해 scrollDirection, scrollY util 추가
